pyproject.toml
[project]
name = "task-management"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-filter>=24.3",
    "django==4.2",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "drf-yasg>=1.21.8",
    "pyproject-toml>=0.1.0",
    "python-dotenv>=1.0.1",
    "psycopg2-binary>=2.9.10",
]

[tool.isort]
profile = "black"
line_length = 90
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
combine_as_imports = true
